+++ R/covmatrix.R
@@ -88,8 +88,9 @@ covmatrix.splm <- function(object, newdata, cov_type, ...) {
       }
     }
 
-    if (inherits(object$newdata, "sf")) {
-      object$obdata <- sf_to_df(object$newdata)
+    if (inherits(newdata, "sf")) {
+      newdata <- suppressWarnings(sf::st_centroid(newdata))
+      object$obdata <- sf_to_df(newdata)
     } else {
       object$obdata <- newdata
     }